+++ operators.go
@@ -1,8 +1,8 @@
 package main
 
 type NormalOperator struct {
-	Num1 int
-	Num2 int
+	Num1 float64
+	Num2 float64
 }
 
 func (op *NormalOperator) Add() (float64, string, error) {
